+++ weather-app/app/build.gradle
@@ -31,4 +31,5 @@ dependencies {
     implementation 'com.android.support:design:27.+'
     implementation 'com.android.support:support-v4:27.+'
     testImplementation 'junit:junit:4.12'
+    implementation 'android.arch.lifecycle:extensions:1.1.1'
 }
