+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/model/FindResult.java
@@ -1,0 +1,11 @@
+package com.rperazzo.weatherapp.model;
+
+import java.util.List;
+
+public class FindResult {
+    public final List<City> list;
+
+    public FindResult(List<City> list) {
+        this.list = list;
+    }
+}
