+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/services/WeatherService.java
@@ -1,0 +1,18 @@
+package com.rperazzo.weatherapp.services;
+
+import com.rperazzo.weatherapp.model.FindResult;
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+import retrofit2.http.Query;
+
+public interface WeatherService {
+
+    @GET("find")
+    Call<FindResult> find(
+            @Query("q") String cityName,
+            @Query("units") String units,
+            @Query("appid") String apiKey
+    );
+
+}
