+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/utils/Const.java
@@ -1,0 +1,13 @@
+package com.rperazzo.weatherapp.utils;
+
+public class Const {
+
+    public static final String TEMPERATURE_UNIT_KEY = "TEMPERATURE_UNIT_KEY";
+
+    public static final String PREFERENCE_NAME = "com.rperazzo.weatherapp.shared";
+
+    public static final String APP_KEY = "520d6b47a12735bee8f69c57737d145f";
+
+    public static final String API_URL = "http://api.openweathermap.org/data/2.5/";
+
+}
