+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/utils/WeatherManager.java
@@ -1,0 +1,28 @@
+package com.rperazzo.weatherapp.utils;
+
+import com.rperazzo.weatherapp.services.WeatherService;
+
+import okhttp3.OkHttpClient;
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+public class WeatherManager {
+
+    private static OkHttpClient mClient = new OkHttpClient();
+
+    public static WeatherService getService() {
+        Retrofit retrofit = new Retrofit.Builder()
+                .baseUrl(Const.API_URL)
+                .addConverterFactory(GsonConverterFactory.create())
+                .client(mClient)
+                .build();
+
+        return retrofit.create(WeatherService.class);
+    }
+
+
+
+
+}
+
+
