+++ weather-app/app/src/main/res/menu/main_menu.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
     <item
         android:id="@+id/menu_celcius"
         app:showAsAction="always"
